requirements.txt
Django==1.11.11
djangorestframework==3.9.0
numpy==1.15.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pytz==2018.7
